реализация по умолчанию метода базового интерфейса вью валидация данных добавления рабочего помещения вывод ошибок валидации

+++ app/src/main/java/com/example/healthyapp/repo/WorkplaceRepo.kt
@@ -42,7 +42,7 @@ interface WorkplaceRepo {
     fun saveRoom(
         placement: Placement,
         onSuccess: () -> Unit,
-        onError: () -> Unit
+        onError: (Int) -> Unit
     )
 
 }
